[schemaregistry] Move performance tests to service level
