--- sdk/schemaregistry/perf-tests/schema-registry-avro/package.json
@@ -1,79 +1,0 @@
-{
-  "name": "@azure-tests/perf-schema-registry-avro",
-  "sdk-type": "perf-test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./dist/commonjs/index.js",
-  "module": "./dist/esm/index.js",
-  "keywords": [],
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "dependencies": {
-    "@azure-tools/test-perf": "^1.0.0",
-    "@azure/identity": "^4.5.0",
-    "@azure/schema-registry": "1.3.0",
-    "@azure/schema-registry-avro": "^1.1.0",
-    "dotenv": "^16.0.0",
-    "tslib": "^2.8.1"
-  },
-  "devDependencies": {
-    "@azure/dev-tool": "^1.0.0",
-    "@types/node": "^18.0.0",
-    "eslint": "^9.9.0",
-    "typescript": "~5.8.2"
-  },
-  "private": true,
-  "scripts": {
-    "build": "npm run clean && dev-tool run build-package",
-    "build:samples": "echo skipped",
-    "build:test": "echo skipped",
-    "check-format": "dev-tool run vendored prettier --list-different --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"src/**/*.ts\" \"*.{js,json}\"",
-    "clean": "dev-tool run vendored rimraf --glob dist dist-* types *.tgz *.log",
-    "format": "dev-tool run vendored prettier --write --config ../../../../.prettierrc.json --ignore-path ../../../../.prettierignore \"src/**/*.ts\" \"*.{js,json}\"",
-    "integration-test": "echo skipped",
-    "integration-test:browser": "echo skipped",
-    "integration-test:node": "echo skipped",
-    "lint": "eslint -c ../../../../common/tools/eslint-plugin-azure-sdk/eslint.perftests.config.mjs src",
-    "lint:fix": "eslint -c ../../../../common/tools/eslint-plugin-azure-sdk/eslint.perftests.config.mjs src --fix --fix-type [problem,suggestion]",
-    "pack": "npm pack 2>&1",
-    "perf-test:node": "npm run build && node dist/esm/index.js",
-    "test": "echo skipped",
-    "test:browser": "echo skipped",
-    "test:node": "echo skipped",
-    "unit-test": "echo skipped",
-    "unit-test:browser": "echo skipped",
-    "unit-test:node": "echo skipped",
-    "update-snippets": "echo skipped"
-  },
-  "type": "module",
-  "tshy": {
-    "exports": {
-      "./package.json": "./package.json",
-      ".": "./src/index.ts"
-    },
-    "dialects": [
-      "esm",
-      "commonjs"
-    ],
-    "selfLink": false
-  },
-  "files": [
-    "dist/",
-    "README.md",
-    "LICENSE"
-  ],
-  "exports": {
-    "./package.json": "./package.json",
-    ".": {
-      "import": {
-        "types": "./dist/esm/index.d.ts",
-        "default": "./dist/esm/index.js"
-      },
-      "require": {
-        "types": "./dist/commonjs/index.d.ts",
-        "default": "./dist/commonjs/index.js"
-      }
-    }
-  },
-  "types": "./dist/commonjs/index.d.ts"
-}
