--- sdk/schemaregistry/perf-tests/schema-registry-avro/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "extends": "../../../../tsconfig",
-  "compilerOptions": {
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "rootDir": ".",
-    "paths": {
-      "@azure-tests/perf-schema-registry-avro": ["./src/index"]
-    }
-  },
-  "compileOnSave": true,
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.mts",
-    "src/**/*.cts",
-    "samples-dev/**/*.ts",
-    "aex/**/*.ts",
-    "aex/**/*.mts",
-    "aex/**/*.cts"
-  ]
-}
