+++ sdk/schemaregistry/schema-registry-avro-perf-tests/src/avroSerializerTest.spec.ts
@@ -4,8 +4,8 @@
 import { PerfTest, getEnvVar } from "@azure-tools/test-perf";
 import { AvroSerializer } from "@azure/schema-registry-avro";
 import { SchemaRegistryClient } from "@azure/schema-registry";
-import "dotenv/config";
 import { DefaultAzureCredential } from "@azure/identity";
+import "dotenv/config";
 
 export abstract class AvroSerializerTest<TOptions> extends PerfTest<TOptions> {
   static schema = JSON.stringify({
