+++ sdk/schemaregistry/schema-registry-avro-perf-tests/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "references": [
+    {
+      "path": "./tsconfig.src.json"
+    }
+  ]
+}
