+++ sdk/schemaregistry/schema-registry-avro-perf-tests/tsconfig.src.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../tsconfig.lib.json"
+}
